Use a list of strings instead of IFolders for X-GM-LABELS
The FolderCache will not necessarily have a mapping of all of the
user's folders, so using this will be unreliable. We also can't
send new commands to lookup/get the corresponding folders because
the command loop is not reentrant.

The easiest solution is to simply use a list of strings and have the
app deal with that.