Refactored SmtpClient, Pop3Client, and ImapClient to share a base class
MessageService implements IMessageService and is the new base class
for SmtpClient, Pop3Client, and ImapClient.

It is the new home for the X509 client certificates and also
adds some new Connect() and Authenticate() wrappers that might
be a little more obvious to developers coming to MailKit from
other mail libraries that don't use Uris or ICredentials.