Marked POP3 methods that take uids as [Obsolete]
IMHO, I made a bad design decision to offer these APIs in the first place
since it requires the Pop3Client to maintain a UID-to-SequenceID mapping
that (hopefully) the consumer of the MailKit library is already doing
themselves and is thus unnecessarily duplicating what could be a large
amount of data.

There's also the problem of developers expecting to simply be able to
connect to a POP3 server and immediately use one of these methods
before calling GetMessageUids() (which is the only way for the
Pop3Client to obtain the UID mapping). This commit does work around
this problem by adding logic to each method override taking a uid
(or list of uids) to internally call GetMessageUids(), but this is
not an ideal situation.