Always return UniqueId(s) for folder.Append(), CopyTo() and MoveTo() if we can
Now that UniqueId has a Validity value, we can leave it up to the caller
to make sure that the UidValidity of the destination folder matches the
Validity value of the UniqueId.

Previously, the destination folder would have had to have been opened
at some point prior to the append/copy/move operation so that the
UidValidity property of the folder would be known.

This is no longer necessary.