Added a MessageSummaryFetched event to IMailFolder
This allows developers to get immediate feedback as message summaries
are fetched from, for example, an IMAP server rather than having to
wait for the IMailFolder.Fetch() method to complete. It also means
that if the connection to the IMAP server drops, the client will not
have to start over from scratch (assuming it is listening to the
MessageSummaryFetched event and using it to do book keeping).