Only trim CR and LF from the end of POP3 response lines
This is to prevent potential problems with multi-line responses
where a line might be something like ".<wsp>\r\n" which should
NOT be treated as a termination sequence.

Since these cases are only for simple multi-line commands such as
LIST, UIDL, LANG and CAPA, it's not likely this ever would have
been an issue, but might as well robustify the code anyway.