Refactored ImapFolder's FETCH handlers to reduce duplication
This will also have the added bonus of allowing NOTIFY support
to work properly if the user elects to be notified of new
messages and requesting MessageSummaryItems other than FLAGS,
X-GM-LABELS, and/or MODSEQ.

Note: we also now call the virtual OnMessageSummaryFetched()
method with the returned IMessageSummary. For clients that
do not override this method, it will also manifest itself
in a MessageSummaryFetched event being fired.