Improved FetchSummaryItemsAsync() BODY parsing logic
Keep track of whether or not we are reading the HEADER.FIELDS
list and, if so, check whether the REFERENCES header is listed.

If it is listed, only then do we set IMessageSummary.Fields
to include the MessageSummaryItems.References value.

Also always set IMessageSummary.Fields to include the
MessageSummaryItems.Headers value if we get any headers.