Refactored ParseFolderListAsync() to not create a new ImapFolder unless it needs to
We want to make sure not to create a new ImapFolder instance until
after we've parsed any LIST extensions (for "OLDNAME") so that
if anyone subclasses ImapFolder, they don't get confused by us
temporarily creating a new ImapFolder and then updating an old
ImapFolder to have the name of the newly created ImapFolder.

Also refactored ImapFolder/MailFolder a bit so that ImapUtils
didn't have to set protected properties on MailFolder.