Updated SmtpClient to measure the message size in more cases
If the ITransferProgress is specified, always calculate the size of
the message for progress reporting purposes.

If the SMTP server supports the SIZE extension, get the size so
that we can send the SIZE= parameter in the MAIL FROM command.

Note: Previously, the size was only claculated as part of the BDAT
command.

Also: Provide a way for subclasses to override the size calculation
so that if they have the information cached, they can return that
value rather than needing to actually measure the output of
MimeMessage.WriteTo().

Implements the solution proposed in issue #896