Modified NetworkStream.ReadAsync() to poll instead of using Socket.ReceiveAsync()
The problem with ReceiveAsync() is that the only way to cancel that is
to close the socket. Otherwise what happens is that the async operation
continues even though we break out of the await, so when the IMAP client
eventually goes to ReadAsync() again, it could get an
OperationInvalidException due to the "cancelled" ReadAsync() still executing.

The custom SslStream.ReadAsync() has also been dropped because it opens
another can of worms. It is possible for the base SslStream to have the
remainder of the command response buffered. If that is the case, then
waiting for more data to arrive on the underlying Socket will cause a
hang (since it will never arrive).