Switch to using tls-server-end-point for NTLM channel-bindings
Based on https://en.hackndo.com/ntlm-relay/ the channel-binding is
a hash of the TLS certificate which suggests that NTLM uses the
tls-server-end-point channel-binding.

This may technically still be wrong because the NTLM specifications
say that the "unhashed" channel-binding data needs to be hashed using
MD5. The "tls-server-end-point" channel-binding data is already hashed
using the certificate's specified hash algorithm (SHA-256 in most cases).

Do we need to MD5 hash the SHA-256 hashed certificate data? Or do we need
to MD5 hash the SHA-256 hashed data?

Gotta love vague documentation.