Optimize FetchSummaryContext.BinarySearch (used by Add) for the common case
In most situations, untagged FETCH responses will be in sequential order
and not random, so we can quickly check to see if the new index is larger
than the last index in the list - if it is, great, no need to binary search
for the correct location to insert the new FETCH response.

For a test GMail folder containing ~50k messages, this improves performance
from ~32s down to ~26s to FETCH the entire folder contents.