Improve the exception message in case of certificate name mismatch
Before this commit, the message was:
> An error occurred while attempting to establish an SSL or TLS connection.
>
> The host name did not match the name given in the server's SSL certificate.

After this commit, the message becomes:
> An error occurred while attempting to establish an SSL or TLS connection.
> 
> The host name (smtp.contoso.com) did not match any of the names given in the server's SSL certificate:
>   • *.mail.eo.outlook.com
>   • *.mail.protection.outlook.com
>   • *.mail.protection.outlook.de
>   • *.olc.protection.outlook.com
>   • *.pamx1.hotmail.com
>   • mail.messaging.microsoft.com
>   • mail.protection.outlook.com
>   • outlook.com

Providing all the certificate subject alternative names directly in the error message will greatly help diagnosing certificate name mismatch issues.

It would have helped me to understand that `smtp.contoso.com` is an alias for `contoso-com.mail.protection.outlook.com` and that using the latter instead of the former solves the connection issue.