Updated ImapStream, ImapEngine, and ImapCommand to reuse ByteArrayBuilders
This drastically reduces the number of allocations made when tokenizing
IMAP responses.

ImapCommand's usage was not really a major issue, but since ImapEngine.ReadLine/Async()
needed a reusable ByteArrayBuilder anyway, might as well share that with ImapCommand.

Also modified ByteArrayBuilder.Clear() to downsize its buffer if it has grown too large.

One potential problem with reusing ByteArrayBuilders is that, because they can
grow for some abnormally long tokens/lines/commands/etc, those oversized buffers
will remain referenced by the ImapStream/ImapEngine until they are disposed which
could be the life of the program.

If we opportunistically scale back the size of the buffers when they are Clear()'d,
then we can periodically reduce memory usage and allow those larger buffers to be
used elsewhere.