docs: clarify/add the recommendation to set AutoSubmitted if needed
IMHO, many people using this lib may use it to automatically send replies. Sure, some may have some human in the loop, but we can assume that.

As such, IMHO; the FAQ should point that out.

Related to https://github.com/jstedfast/MimeKit/issues/938 of course and there are plenty of more headers to set in such a case, but it feels like this was the most standardized one and in any case the minimal one to do?

Also, the FAQ entry is about replying, so the header value does exactly reflect that. If you want to suppress responses or so, that may be an additional requirement.